requirements.txt
beautifulsoup4
Bottleneck
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
distro
et-xmlfile
numexpr
numpy==1.26.4
openpyxl
pandas==2.1.1
pdfminer.six==20231228
pdfplumber==0.11.0
Pillow==10.2.0
pycparser==2.21
pypdfium2==4.28.0
python-dateutil
pytz
setuptools==68.2.2
six
soupsieve
tabula-py==2.7.0
tzdata
wheel==0.41.2
matplotlib==3.8.3
